hadith-data/Postgres_Database_setup/database_config.py
"""
Database configuration for PostgreSQL connection
"""
import os
from sqlalchemy import create_engine
from sqlalchemy.orm import sessionmaker
from sqlalchemy.ext.declarative import declarative_base
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Database configuration
DATABASE_URL = os.getenv('DATABASE_URL', 'postgresql://postgres:ranaatif@localhost:5432/hadith_db')

# Create engine
engine = create_engine(DATABASE_URL, echo=False)  # Disabled SQL logging for cleaner output

# Create session
SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)

# Base class for models
Base = declarative_base()

def get_db():
    """Get database session"""
    db = SessionLocal()
    try:
        yield db
    finally:
        db.close()
