requirements.txt
altair==5.5.0
annotated-types==0.7.0
anthropic==0.57.1
anyio==4.9.0
arabert==1.0.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.2.1
cachetools==5.5.2
camel-tools==1.5.6
catalogue==2.0.10
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
comm==0.2.2
confection==0.1.5
cryptography==45.0.5
cymem==2.0.11
debugpy==1.8.12
decorator==5.1.1
dill==0.4.0
distro==1.9.0
docopt==0.6.2
editdistance==0.8.1
emoji==1.4.2
exceptiongroup==1.3.0
executing==2.2.0
farasapy==0.0.14
fastapi==0.116.1
filelock==3.18.0
filetype==1.2.0
fsspec==2025.7.0
future==1.0.0
git-filter-repo==2.47.0
gitdb==4.0.12
gitpython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.33.4
hyperframe==6.1.0
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-client==8.6.3
jupyter-core==5.7.2
langchain==0.3.26
langchain-anthropic==0.3.17
langchain-core==0.3.69
langchain-google-genai==2.0.10
langchain-huggingface==0.3.0
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.4.6
language-data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
muddler==0.1.3
murmurhash==1.0.13
narwhals==1.47.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
openai==1.96.1
orjson==3.11.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
pdfminer-six==20250506
pdfplumber==0.11.7
pillow==11.3.0
platformdirs==4.3.6
portalocker==2.10.1
preshed==3.0.10
prompt-toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.1.1
pure-eval==0.2.3
pyarabic==0.6.15
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydeck==0.9.1
pygments==2.19.2
pyparsing==3.2.3
pypdfium2==4.30.1
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==308
pyyaml==6.0.2
pyzmq==26.2.0
qdrant-client==1.14.3
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==5.0.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlalchemy==2.0.41
srsly==2.5.1
stack-data==0.6.3
starlette==0.47.1
streamlit==1.46.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.19.1
toml==0.10.2
torch==2.7.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.43.4
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.17.2
zstandard==0.23.0
